requirements.txt
Django>=1.8
django-cms>=3.0
djangorestframework>=3.5.0
django-appconf>=1.0.1
requests>=2.0.0
